fix: simplify workflows view
